requirements.txt
aniso8601==7.0.0
click==7.1.2
Flask==1.1.4
Flask-GraphQL==2.0.1
Flask-Script==2.0.6
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
haversine==2.8.0
ijson==3.2.0.post0
iso8601==1.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==2.0.1
promise==2.3
psycopg2-binary==2.9.6
Rx==1.6.3
six==1.16.0
Werkzeug==1.0.1